Create a parallel thread for processing AzureKinect camera input
